Ceres Optimization
has_qr is used to see whether there is qr_msg. If no qr_msg, the variable from keyframe will be set to constant.
We create variable for keyframe origin and update it also. That's why there are two if functions are modified.
An qr edge has been added using the relative qr information in the keyframe structure.